Updated server.js
Updated server js so that posting to home with the username will redirect to the app.get section in accordance with the correct username. This will allow user to successfully navigate to their homepage after logging in. Also added header to all functions in relation to the userSchema in the mongoose.js file to separate the commands.