Added sliding feature for group page
Added a feature that allows a screen to slide onto the page from underneath the footer, pushing the footer up alongside it. The page that is pushed up from pressing the button will hold a users top recommended songs. They can use these songs or search songs to suggest songs to add to queue.