Update on adding user data to UI
Working on having the webpage automatically upload up to 50 top songs for the user to choose from to suggest to the queue. First major step was just finished and refined in this commit - turning the users authorization code into an access key. Also added some basic CSS to format the songs for the time being.