Added fetch users top songs
Made it so that webpage can fetch the users top songs from their spotify account. Soon to make it so that if no top songs are present, users saved songs will appear instead. Along with this, soon to make it so that user can search for any song on spotify.