Changed User Schema
Made it so that a user object will now hold a spotify refresh token as well as an access token. After the spotify access token expires (this takes 3600 seconds), spotify api will return a 401 response. Next to be implemented is interpreting this 401, refreshing the access_token, and trying the desired call again with a renewed token.