Added queue page
Added a page to view the groups current queue. Depending on if formatting changes, this will turn into the page to view the groups suggestions. This page will pull from the suggestions stored in mongoDB and will put a song from the queue/ suggestion from the group on the screen. If you click on the queue/suggestion bar, the page will slide onto the page from the right. It will slide off the of the screen in the opposite fashion when the user is done with it.