Updates
Updates to many aspects of the code have been made. Working on making this project in good condition for submission for CS361's final project assignment. All functionalities involving the group page (other than invite to group) work. With this, further styling will need to be completed. And code will have to be cleaned up.